CHANGES IN BUTTON COMPONENT
Extracted themes into a separate file,
added onClick prop and removed classNames from
ButtonPage and instead changed all this system
into just passing props.
Added variants primary, secondary and danger,
which means no longer the component depends
on text that was written as children of the
button in the ButtonPage (now we use a variant
instead).